# Empresa-OpenApi.yaml
openapi: 3.0.0
info:
  title: Cuentas
  version: '1.0'
  description: Operaciones con Cuentas
  contact:
    name: William
    url: 'http://wquimis.com'
    email: wquimis@gmail.com
  termsOfService: 'http://wquimis.com'
servers:
  - url: 'http://localhost:3000'
...
paths:
  /clientes:
    parameters: []
    post:
      summary: Crear Cliente
      operationId: crearcliente
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
              examples:
                cliente:
                  value:
                    idCliente: 0
                    nombres: string
                    genero: M
                    edad: 17
                    identificacion: '0900000000'
                    direccion: string
                    telefono: '04000000'
                    clave: pa$$word
                    estado: A
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
            examples:
              cliente:
                value:
                  idCliente: 0
                  nombres: string
                  genero: M
                  edad: 17
                  identificacion: '0900000000'
                  direccion: string
                  telefono: '04000000'
                  clave: pa$$word
                  estado: A
        description: Devuelve el objeto creado con su ID
      description: Creacion de Clientes
      tags:
        - Clientes
    get:
      summary: Consultar Clientes
      operationId: obtenercliente
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
              examples:
                cliente:
                  value:
                    - idCliente: 0
                      nombres: string
                      genero: M
                      edad: 17
                      identificacion: '0900000000'
                      direccion: string
                      telefono: '04000000'
                      clave: pa$$word
                      estado: A
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      description: Consulta de Clientes
      parameters:
        - schema:
            type: integer
          in: query
          name: idcliente
          description: Identificador de cliente
      tags:
        - Clientes
  '/clientes/{idcliente}':
    put:
      summary: Actualizar Cliente
      operationId: actualizarcliente
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
              examples:
                cliente:
                  value:
                    idCliente: 0
                    nombres: string
                    genero: M
                    edad: 17
                    identificacion: '0900000000'
                    direccion: string
                    telefono: '04000000'
                    clave: pa$$word
                    estado: A
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
            examples:
              cliente:
                value:
                  idCliente: 0
                  nombres: string
                  genero: M
                  edad: 17
                  identificacion: '0900000000'
                  direccion: string
                  telefono: '04000000'
                  clave: pa$$word
                  estado: A
          application/xml:
            schema:
              type: object
              properties: {}
          multipart/form-data:
            schema:
              type: object
              properties: {}
        description: Devuelve el cliente actualizado
      description: Actualizacion de clientes
      tags:
        - Clientes
    parameters:
      - schema:
          type: integer
        name: idcliente
        in: path
        required: true
    delete:
      summary: Eliminar Cliente
      operationId: eleminarcliente
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      description: Elimina cliente
      tags:
        - Clientes
...
components:
  schemas:
...
    Error:
      title: Error
      type: object
      x-examples: {}
      description: Modelos de errores
      properties:
        idError:
          type: integer
        mensajeTecnico:
          type: string
        mensajeUsuario:
          type: string
        nivelError:
          type: string
    Cliente:
      type: object
      x-examples: {}
      title: Cliente
      description: Modelo Cliente
      properties:
        idCliente:
          type: integer
        nombres:
          type: string
          maxLength: 100
        genero:
          type: string
          maxLength: 1
          default: M
        edad:
          type: integer
          minimum: 1
          maximum: 150
          default: 17
        identificacion:
          type: string
          maxLength: 10
          default: '0900000000'
        direccion:
          type: string
          maxLength: 200
        telefono:
          type: string
          maxLength: 10
          default: '04000000'
        clave:
          type: string
          maxLength: 20
          format: password
        estado:
          type: string
          maxLength: 1
          default: A
      required:
        - nombres
        - identificacion
        - direccion
        - telefono
        - clave
...