# EmpresaAPI/Models/Error.cs
/*
 * Cuentas
 *
 * Operaciones con Cuentas
 *
 * OpenAPI spec version: 1.0
 * Contact: wquimis@gmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace EmpresaAPI.Models
{ 
    /// <summary>
    /// Modelos de errores
    /// </summary>
    [DataContract]
    public partial class Error : IEquatable<Error>
    { 
        /// <summary>
        /// Gets or Sets IdError
        /// </summary>

        [DataMember(Name="idError")]
        public int? IdError { get; set; }

        /// <summary>
        /// Gets or Sets MensajeTecnico
        /// </summary>

        [DataMember(Name="mensajeTecnico")]
        public string MensajeTecnico { get; set; }

        /// <summary>
        /// Gets or Sets MensajeUsuario
        /// </summary>

        [DataMember(Name="mensajeUsuario")]
        public string MensajeUsuario { get; set; }

        /// <summary>
        /// Gets or Sets NivelError
        /// </summary>

        [DataMember(Name="nivelError")]
        public string NivelError { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Error {\n");
            sb.Append("  IdError: ").Append(IdError).Append("\n");
            sb.Append("  MensajeTecnico: ").Append(MensajeTecnico).Append("\n");
            sb.Append("  MensajeUsuario: ").Append(MensajeUsuario).Append("\n");
            sb.Append("  NivelError: ").Append(NivelError).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Error)obj);
        }

        /// <summary>
        /// Returns true if Error instances are equal
        /// </summary>
        /// <param name="other">Instance of Error to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Error other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    IdError == other.IdError ||
                    IdError != null &&
                    IdError.Equals(other.IdError)
                ) && 
                (
                    MensajeTecnico == other.MensajeTecnico ||
                    MensajeTecnico != null &&
                    MensajeTecnico.Equals(other.MensajeTecnico)
                ) && 
                (
                    MensajeUsuario == other.MensajeUsuario ||
                    MensajeUsuario != null &&
                    MensajeUsuario.Equals(other.MensajeUsuario)
                ) && 
                (
                    NivelError == other.NivelError ||
                    NivelError != null &&
                    NivelError.Equals(other.NivelError)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (IdError != null)
                    hashCode = hashCode * 59 + IdError.GetHashCode();
                    if (MensajeTecnico != null)
                    hashCode = hashCode * 59 + MensajeTecnico.GetHashCode();
                    if (MensajeUsuario != null)
                    hashCode = hashCode * 59 + MensajeUsuario.GetHashCode();
                    if (NivelError != null)
                    hashCode = hashCode * 59 + NivelError.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Error left, Error right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Error left, Error right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}